requirements.txt
Django~=4.2.4
djangorestframework~=3.14.0
matplotlib==3.7.2
pandas==2.0.3
drf-spectacular==0.26.4